Re-organise api functions and hooks, remove what is not used
